splitting decoder into 32/64 to solve macOS failing tests.
