+++ src/core/coding/ojph_block_decoder64.cpp
@@ -287,145 +287,6 @@ namespace ojph {
     //************************************************************************/
     /** @brief Read and unstuff data from a backwardly-growing segment
      *
-     *  This reader can read up to 8 bytes from before the VLC segment.
-     *  Care must be taken not read from unreadable memory, causing a 
-     *  segmentation fault.
-     * 
-     *  Note that there is another subroutine rev_read_mrp that is slightly
-     *  different.  The other one fills zeros when the buffer is exhausted.
-     *  This one basically does not care if the bytes are consumed, because
-     *  any extra data should not be used in the actual decoding.
-     *
-     *  Unstuffing is needed to prevent sequences more than 0xFF8F from 
-     *  appearing in the bits stream; since we are reading backward, we keep
-     *  watch when a value larger than 0x8F appears in the bitstream. 
-     *  If the byte following this is 0x7F, we unstuff this byte (ignore the 
-     *  MSB of that byte, which should be 0).
-     *
-     *  @param [in]  vlcp is a pointer to rev_struct structure
-     */
-    static inline 
-    void rev_read(rev_struct *vlcp)
-    {
-      //process 4 bytes at a time
-      if (vlcp->bits > 32)  // if there are more than 32 bits in tmp, then 
-        return;             // reading 32 bits can overflow vlcp->tmp
-      ui32 val = 0;
-      //the next line (the if statement) needs to be tested first
-      if (vlcp->size > 3)  // if there are more than 3 bytes left in VLC
-      {
-        // (vlcp->data - 3) move pointer back to read 32 bits at once
-        val = *(ui32*)(vlcp->data - 3); // then read 32 bits
-        vlcp->data -= 4;          // move data pointer back by 4
-        vlcp->size -= 4;          // reduce available byte by 4
-      }
-      else if (vlcp->size > 0)
-      { // 4 or less
-        int i = 24;
-        while (vlcp->size > 0) {   
-          ui32 v = *vlcp->data--; // read one byte at a time
-          val |= (v << i);        // put byte in its correct location
-          --vlcp->size;
-          i -= 8;
-        }
-      }
-
-      //accumulate in tmp, number of bits in tmp are stored in bits
-      ui32 tmp = val >> 24;  //start with the MSB byte
-      ui32 bits;
-
-      // test unstuff (previous byte is >0x8F), and this byte is 0x7F
-      bits = 8 - ((vlcp->unstuff && (((val >> 24) & 0x7F) == 0x7F)) ? 1 : 0);
-      bool unstuff = (val >> 24) > 0x8F; //this is for the next byte
-
-      tmp |= ((val >> 16) & 0xFF) << bits; //process the next byte
-      bits += 8 - ((unstuff && (((val >> 16) & 0x7F) == 0x7F)) ? 1 : 0);
-      unstuff = ((val >> 16) & 0xFF) > 0x8F;
-
-      tmp |= ((val >> 8) & 0xFF) << bits;
-      bits += 8 - ((unstuff && (((val >> 8) & 0x7F) == 0x7F)) ? 1 : 0);
-      unstuff = ((val >> 8) & 0xFF) > 0x8F;
-
-      tmp |= (val & 0xFF) << bits;
-      bits += 8 - ((unstuff && ((val & 0x7F) == 0x7F)) ? 1 : 0);
-      unstuff = (val & 0xFF) > 0x8F;
-
-      // now move the read and unstuffed bits into vlcp->tmp
-      vlcp->tmp |= (ui64)tmp << vlcp->bits;
-      vlcp->bits += bits;
-      vlcp->unstuff = unstuff; // this for the next read
-    }
-
-    //************************************************************************/
-    /** @brief Initiates the rev_struct structure and reads a few bytes to 
-     *         move the read address to multiple of 4
-     *
-     *  There is another similar rev_init_mrp subroutine.  The difference is
-     *  that this one, rev_init, discards the first 12 bits (they have the
-     *  sum of the lengths of VLC and MEL segments), and first unstuff depends
-     *  on first 4 bits.
-     *
-     *  @param [in]  vlcp is a pointer to rev_struct structure
-     *  @param [in]  data is a pointer to byte at the start of the cleanup pass
-     *  @param [in]  lcup is the length of MagSgn+MEL+VLC segments
-     *  @param [in]  scup is the length of MEL+VLC segments
-     */
-    static inline 
-    void rev_init(rev_struct *vlcp, ui8* data, int lcup, int scup)
-    {
-      //first byte has only the upper 4 bits
-      vlcp->data = data + lcup - 2;
-
-      //size can not be larger than this, in fact it should be smaller
-      vlcp->size = scup - 2;
-
-      ui32 d = *vlcp->data--; // read one byte (this is a half byte)
-      vlcp->tmp = d >> 4;    // both initialize and set
-      vlcp->bits = 4 - ((vlcp->tmp & 7) == 7); //check standard
-      vlcp->unstuff = (d | 0xF) > 0x8F; //this is useful for the next byte
-
-      //This code is designed for an architecture that read address should
-      // align to the read size (address multiple of 4 if read size is 4)
-      //These few lines take care of the case where data is not at a multiple
-      // of 4 boundary. It reads 1,2,3 up to 4 bytes from the VLC bitstream.
-      // To read 32 bits, read from (vlcp->data - 3)
-      int num = 1 + (int)(intptr_t(vlcp->data) & 0x3);
-      int tnum = num < vlcp->size ? num : vlcp->size;
-      for (int i = 0; i < tnum; ++i) {
-        ui64 d;
-        d = *vlcp->data--;  // read one byte and move read pointer
-        //check if the last byte was >0x8F (unstuff == true) and this is 0x7F
-        ui32 d_bits = 8 - ((vlcp->unstuff && ((d & 0x7F) == 0x7F)) ? 1 : 0);
-        vlcp->tmp |= d << vlcp->bits; // move data to vlcp->tmp
-        vlcp->bits += d_bits;
-        vlcp->unstuff = d > 0x8F; // for next byte
-      }
-      vlcp->size -= tnum;
-      rev_read(vlcp);  // read another 32 buts
-    }
-
-    //************************************************************************/
-    /** @brief Retrieves 32 bits from the head of a rev_struct structure 
-     *
-     *  By the end of this call, vlcp->tmp must have no less than 33 bits
-     *
-     *  @param [in]  vlcp is a pointer to rev_struct structure
-     */
-    static inline 
-    ui32 rev_fetch(rev_struct *vlcp)
-    {
-      if (vlcp->bits < 32)  // if there are less then 32 bits, read more
-      {
-        rev_read(vlcp);     // read 32 bits, but unstuffing might reduce this
-        if (vlcp->bits < 32)// if there is still space in vlcp->tmp for 32 bits
-          rev_read(vlcp);   // read another 32
-      }
-      return (ui32)vlcp->tmp; // return the head (bottom-most) of vlcp->tmp
-    }
-
-    //************************************************************************/
-    /** @brief Read and unstuff data from a backwardly-growing segment
-     *
      *  This reader reads 8 bits from the VLC segment. It fills zeros when 
      *  the buffer is exhausted; we basically do not care about these zeros 
      *  because we should not need them -- any extra data should not be used 
@@ -513,21 +374,6 @@ namespace ojph {
         rev_read8(vlcp); // read 8 bits, but unstuffing might reduce this
       return vlcp->tmp;  // return unstuff decoded bits
     }    
-
-    //************************************************************************/
-    /** @brief Consumes num_bits from a rev_struct structure
-     *
-     *  @param [in]  vlcp is a pointer to rev_struct structure
-     *  @param [in]  num_bits is the number of bits to be removed
-     */
-    static inline 
-    ui32 rev_advance(rev_struct *vlcp, ui32 num_bits)
-    {
-      assert(num_bits <= vlcp->bits); // vlcp->tmp must have more than num_bits
-      vlcp->tmp >>= num_bits;         // remove bits
-      vlcp->bits -= num_bits;         // decrement the number of bits
-      return (ui32)vlcp->tmp;
-    }
 
     //************************************************************************/
     /** @brief Consumes num_bits from a rev_struct structure
@@ -899,897 +745,6 @@ namespace ojph {
         frwd_read8<X>(msp);
       return msp->tmp;
     }    
-
-    //************************************************************************/
-    /** @brief Decodes one codeblock, processing the cleanup, siginificance
-     *         propagation, and magnitude refinement pass
-     *
-     *  @param [in]   coded_data is a pointer to bitstream
-     *  @param [in]   decoded_data is a pointer to decoded codeblock data buf.
-     *  @param [in]   missing_msbs is the number of missing MSBs
-     *  @param [in]   num_passes is the number of passes: 1 if CUP only,
-     *                2 for CUP+SPP, and 3 for CUP+SPP+MRP
-     *  @param [in]   lengths1 is the length of cleanup pass
-     *  @param [in]   lengths2 is the length of refinement passes (either SPP
-     *                only or SPP+MRP)
-     *  @param [in]   width is the decoded codeblock width 
-     *  @param [in]   height is the decoded codeblock height
-     *  @param [in]   stride is the decoded codeblock buffer stride 
-     *  @param [in]   stripe_causal is true for stripe causal mode
-     */
-    bool ojph_decode_codeblock32(ui8* coded_data, ui32* decoded_data,
-                                 ui32 missing_msbs, ui32 num_passes,
-                                 ui32 lengths1, ui32 lengths2,
-                                 ui32 width, ui32 height, ui32 stride,
-                                 bool stripe_causal)
-    {
-      static bool insufficient_precision = false;
-      static bool modify_code = false;
-      static bool truncate_spp_mrp = false;
-
-      if (num_passes > 1 && lengths2 == 0)
-      {
-        OJPH_WARN(0x00010001, "A malformed codeblock that has more than "
-                              "one coding pass, but zero length for "
-                              "2nd and potential 3rd pass.");
-        num_passes = 1;
-      }
-
-      if (num_passes > 3)
-      {
-        OJPH_WARN(0x00010002, "We do not support more than 3 coding passes; "
-                              "This codeblocks has %d passes.",
-                              num_passes);
-        return false;
-      }
-
-      if (missing_msbs > 30) // p < 0
-      {
-        if (insufficient_precision == false) 
-        {
-          insufficient_precision = true;
-          OJPH_WARN(0x00010003, "32 bits are not enough to decode this "
-                                "codeblock. This message will not be "
-                                "displayed again.");
-        }
-        return false;
-      }       
-      else if (missing_msbs == 30) // p == 0
-      { // not enough precision to decode and set the bin center to 1
-        if (modify_code == false) {
-          modify_code = true;
-          OJPH_WARN(0x00010004, "Not enough precision to decode the cleanup "
-                                "pass. The code can be modified to support "
-                                "this case. This message will not be "
-                                "displayed again.");
-        }
-         return false;         // 32 bits are not enough to decode this
-       }
-      else if (missing_msbs == 29) // if p is 1, then num_passes must be 1
-      {
-        if (num_passes > 1) {
-          num_passes = 1;
-          if (truncate_spp_mrp == false) {
-            truncate_spp_mrp = true;
-            OJPH_WARN(0x00010005, "Not enough precision to decode the SgnProp "
-                                  "nor MagRef passes; both will be skipped. "
-                                  "This message will not be displayed "
-                                  "again.");
-          }
-        }
-      }
-      ui32 p = 30 - missing_msbs; // The least significant bitplane for CUP
-      // There is a way to handle the case of p == 0, but a different path
-      // is required
-
-      if (lengths1 < 2)
-      {
-        OJPH_WARN(0x00010006, "Wrong codeblock length.");
-        return false;
-      }
-
-      // read scup and fix the bytes there
-      int lcup, scup;
-      lcup = (int)lengths1;  // length of CUP
-      //scup is the length of MEL + VLC
-      scup = (((int)coded_data[lcup-1]) << 4) + (coded_data[lcup-2] & 0xF);
-      if (scup < 2 || scup > lcup || scup > 4079) //something is wrong
-        return false;
-
-      // The temporary storage scratch holds two types of data in an 
-      // interleaved fashion. The interleaving allows us to use one
-      // memory pointer.
-      // We have one entry for a decoded VLC code, and one entry for UVLC.
-      // Entries are 16 bits each, corresponding to one quad, 
-      // but since we want to use XMM registers of the SSE family 
-      // of SIMD; we allocated 16 bytes or more per quad row; that is,
-      // the width is no smaller than 16 bytes (or 8 entries), and the
-      // height is 512 quads
-      // Each VLC entry contains, in the following order, starting 
-      // from MSB
-      // e_k (4bits), e_1 (4bits), rho (4bits), useless for step 2 (4bits)
-      // Each entry in UVLC contains u_q
-      // One extra row to handle the case of SPP propagating downwards
-      // when codeblock width is 4
-      ui16 scratch[8 * 513] = {0};       // 8 kB
-
-      // We need an extra two entries (one inf and one u_q) beyond
-      // the last column. 
-      // If the block width is 4 (2 quads), then we use sstr of 8 
-      // (enough for 4 quads). If width is 8 (4 quads) we use 
-      // sstr is 16 (enough for 8 quads). For a width of 16 (8 
-      // quads), we use 24 (enough for 12 quads).
-      ui32 sstr = ((width + 2u) + 7u) & ~7u; // multiples of 8
-
-      ui32 mmsbp2 = missing_msbs + 2;
-
-      // The cleanup pass is decoded in two steps; in step one,
-      // the VLC and MEL segments are decoded, generating a record that 
-      // has 2 bytes per quad. The 2 bytes contain, u, rho, e^1 & e^k.
-      // This information should be sufficient for the next step.
-      // In step 2, we decode the MagSgn segment.
-
-      // step 1 decoding VLC and MEL segments
-      {
-        // init structures
-        dec_mel_st mel;
-        mel_init(&mel, coded_data, lcup, scup);
-        rev_struct vlc;
-        rev_init(&vlc, coded_data, lcup, scup);
-
-        int run = mel_get_run(&mel); // decode runs of events from MEL bitstrm
-                                     // data represented as runs of 0 events
-                                     // See mel_decode description
-
-        ui32 vlc_val;
-        ui32 c_q = 0;
-        ui16 *sp = scratch;
-        //initial quad row
-        for (ui32 x = 0; x < width; sp += 4)
-        {
-          // decode VLC
-          /////////////
-
-          // first quad
-          vlc_val = rev_fetch(&vlc);
-
-          //decode VLC using the context c_q and the head of VLC bitstream
-          ui16 t0 = vlc_tbl0[ c_q + (vlc_val & 0x7F) ];
-
-          // if context is zero, use one MEL event
-          if (c_q == 0) //zero context
-          {
-            run -= 2; //subtract 2, since events number if multiplied by 2
-
-            // Is the run terminated in 1? if so, use decoded VLC code, 
-            // otherwise, discard decoded data, since we will decoded again 
-            // using a different context
-            t0 = (run == -1) ? t0 : 0;
-
-            // is run -1 or -2? this means a run has been consumed
-            if (run < 0) 
-              run = mel_get_run(&mel);  // get another run
-          }
-          //run -= (c_q == 0) ? 2 : 0;
-          //t0 = (c_q != 0 || run == -1) ? t0 : 0;
-          //if (run < 0)
-          //  run = mel_get_run(&mel);  // get another run
-          sp[0] = t0;
-          x += 2;
-
-          // prepare context for the next quad; eqn. 1 in ITU T.814
-          c_q = ((t0 & 0x10U) << 3) | ((t0 & 0xE0U) << 2);
-
-          //remove data from vlc stream (0 bits are removed if vlc is not used)
-          vlc_val = rev_advance(&vlc, t0 & 0x7);
-
-          //second quad
-          ui16 t1 = 0;
-
-          //decode VLC using the context c_q and the head of VLC bitstream
-          t1 = vlc_tbl0[c_q + (vlc_val & 0x7F)]; 
-
-          // if context is zero, use one MEL event
-          if (c_q == 0 && x < width) //zero context
-          {
-            run -= 2; //subtract 2, since events number if multiplied by 2
-
-            // if event is 0, discard decoded t1
-            t1 = (run == -1) ? t1 : 0;
-
-            if (run < 0) // have we consumed all events in a run
-              run = mel_get_run(&mel); // if yes, then get another run
-          }
-          t1 = x < width ? t1 : 0;
-          //run -= (c_q == 0 && x < width) ? 2 : 0;
-          //t1 = (c_q != 0 || run == -1) ? t1 : 0;
-          //if (run < 0)
-          //  run = mel_get_run(&mel);  // get another run
-          sp[2] = t1;
-          x += 2;
-
-          //prepare context for the next quad, eqn. 1 in ITU T.814
-          c_q = ((t1 & 0x10U) << 3) | ((t1 & 0xE0U) << 2);
-
-          //remove data from vlc stream, if qinf is not used, cwdlen is 0
-          vlc_val = rev_advance(&vlc, t1 & 0x7);
-          
-          // decode u
-          /////////////
-          // uvlc_mode is made up of u_offset bits from the quad pair
-          ui32 uvlc_mode = ((t0 & 0x8U) << 3) | ((t1 & 0x8U) << 4);
-          if (uvlc_mode == 0xc0)// if both u_offset are set, get an event from
-          {                     // the MEL run of events
-            run -= 2; //subtract 2, since events number if multiplied by 2
-
-            uvlc_mode += (run == -1) ? 0x40 : 0; // increment uvlc_mode by
-                                                 // is 0x40
-
-            if (run < 0)//if run is consumed (run is -1 or -2), get another run
-              run = mel_get_run(&mel);
-          }
-          //run -= (uvlc_mode == 0xc0) ? 2 : 0;
-          //uvlc_mode += (uvlc_mode == 0xc0 && run == -1) ? 0x40 : 0;
-          //if (run < 0)
-          //  run = mel_get_run(&mel);  // get another run
-
-          //decode uvlc_mode to get u for both quads
-          ui32 uvlc_entry = uvlc_tbl0[uvlc_mode + (vlc_val & 0x3F)];
-          //remove total prefix length
-          vlc_val = rev_advance(&vlc, uvlc_entry & 0x7); 
-          uvlc_entry >>= 3; 
-          //extract suffixes for quad 0 and 1
-          ui32 len = uvlc_entry & 0xF;           //suffix length for 2 quads
-          ui32 tmp = vlc_val & ((1 << len) - 1); //suffix value for 2 quads
-          vlc_val = rev_advance(&vlc, len);
-          uvlc_entry >>= 4;
-          // quad 0 length
-          len = uvlc_entry & 0x7; // quad 0 suffix length
-          uvlc_entry >>= 3;
-          ui16 u_q = (ui16)(1 + (uvlc_entry&7) + (tmp&~(0xFFU<<len)));//kap. 1
-          sp[1] = u_q;
-          u_q = (ui16)(1 + (uvlc_entry >> 3) + (tmp >> len));  //kappa == 1
-          sp[3]= u_q;
-        }
-        sp[0] = sp[1] = 0;
-
-        //non initial quad rows
-        for (ui32 y = 2; y < height; y += 2)
-        {
-          c_q = 0;                                // context
-          ui16 *sp = scratch + (y >> 1) * sstr;   // this row of quads
-
-          for (ui32 x = 0; x < width; sp += 4)
-          {
-            // decode VLC
-            /////////////
-
-            // sigma_q (n, ne, nf)
-            c_q |= ((sp[0 - (si32)sstr] & 0xA0U) << 2);
-            c_q |= ((sp[2 - (si32)sstr] & 0x20U) << 4);
-
-            // first quad
-            vlc_val = rev_fetch(&vlc);
-
-            //decode VLC using the context c_q and the head of VLC bitstream
-            ui16 t0 = vlc_tbl1[ c_q + (vlc_val & 0x7F) ];
-
-            // if context is zero, use one MEL event
-            if (c_q == 0) //zero context
-            {
-              run -= 2; //subtract 2, since events number is multiplied by 2
-
-              // Is the run terminated in 1? if so, use decoded VLC code, 
-              // otherwise, discard decoded data, since we will decoded again 
-              // using a different context
-              t0 = (run == -1) ? t0 : 0;
-
-              // is run -1 or -2? this means a run has been consumed
-              if (run < 0) 
-                run = mel_get_run(&mel);  // get another run
-            }
-            //run -= (c_q == 0) ? 2 : 0;
-            //t0 = (c_q != 0 || run == -1) ? t0 : 0;
-            //if (run < 0)
-            //  run = mel_get_run(&mel);  // get another run
-            sp[0] = t0;
-            x += 2;
-
-            // prepare context for the next quad; eqn. 2 in ITU T.814
-            // sigma_q (w, sw)
-            c_q = ((t0 & 0x40U) << 2) | ((t0 & 0x80U) << 1);
-            // sigma_q (nw)
-            c_q |= sp[0 - (si32)sstr] & 0x80;
-            // sigma_q (n, ne, nf)
-            c_q |= ((sp[2 - (si32)sstr] & 0xA0U) << 2);
-            c_q |= ((sp[4 - (si32)sstr] & 0x20U) << 4);
-
-            //remove data from vlc stream (0 bits are removed if vlc is unused)
-            vlc_val = rev_advance(&vlc, t0 & 0x7);
-
-            //second quad
-            ui16 t1 = 0;
-
-            //decode VLC using the context c_q and the head of VLC bitstream
-            t1 = vlc_tbl1[ c_q + (vlc_val & 0x7F)]; 
-
-            // if context is zero, use one MEL event
-            if (c_q == 0 && x < width) //zero context
-            {
-              run -= 2; //subtract 2, since events number if multiplied by 2
-
-              // if event is 0, discard decoded t1
-              t1 = (run == -1) ? t1 : 0;
-
-              if (run < 0) // have we consumed all events in a run
-                run = mel_get_run(&mel); // if yes, then get another run
-            }
-            t1 = x < width ? t1 : 0;
-            //run -= (c_q == 0 && x < width) ? 2 : 0;
-            //t1 = (c_q != 0 || run == -1) ? t1 : 0;
-            //if (run < 0)
-            //  run = mel_get_run(&mel);  // get another run
-            sp[2] = t1;
-            x += 2;
-
-            // partial c_q, will be completed when we process the next quad
-            // sigma_q (w, sw)
-            c_q = ((t1 & 0x40U) << 2) | ((t1 & 0x80U) << 1);
-            // sigma_q (nw)
-            c_q |= sp[2 - (si32)sstr] & 0x80;
-
-            //remove data from vlc stream, if qinf is not used, cwdlen is 0
-            vlc_val = rev_advance(&vlc, t1 & 0x7);
-          
-            // decode u
-            /////////////
-            // uvlc_mode is made up of u_offset bits from the quad pair
-            ui32 uvlc_mode = ((t0 & 0x8U) << 3) | ((t1 & 0x8U) << 4);
-            ui32 uvlc_entry = uvlc_tbl1[uvlc_mode + (vlc_val & 0x3F)];
-            //remove total prefix length
-            vlc_val = rev_advance(&vlc, uvlc_entry & 0x7);
-            uvlc_entry >>= 3;
-            //extract suffixes for quad 0 and 1
-            ui32 len = uvlc_entry & 0xF;           //suffix length for 2 quads
-            ui32 tmp = vlc_val & ((1 << len) - 1); //suffix value for 2 quads
-            vlc_val = rev_advance(&vlc, len);
-            uvlc_entry >>= 4;
-            // quad 0 length
-            len = uvlc_entry & 0x7; // quad 0 suffix length
-            uvlc_entry >>= 3;
-            ui16 u_q = (ui16)((uvlc_entry & 7) + (tmp & ~(0xFFU << len)));
-            sp[1] = u_q;
-            u_q = (ui16)((uvlc_entry >> 3) + (tmp >> len)); // u_q
-            sp[3] = u_q;
-          }
-          sp[0] = sp[1] = 0;
-        }
-      }
-
-      // step2 we decode magsgn
-      {
-        // We allocate a scratch row for storing v_n values.
-        // We have 512 quads horizontally.
-        // We need an extra entry to handle the case of vp[1]
-        // when vp is at the last column.
-        // Here, we allocate 4 instead of 1 to make the buffer size
-        // a multipled of 16 bytes.
-        const int v_n_size = 512 + 4;
-        ui32 v_n_scratch[v_n_size] = {0};  // 2+ kB
-
-        frwd_struct magsgn;
-        frwd_init<0xFF>(&magsgn, coded_data, lcup - scup);
-
-        ui16 *sp = scratch;
-        ui32 *vp = v_n_scratch;
-        ui32 *dp = decoded_data;
-
-        ui32 prev_v_n = 0;
-        for (ui32 x = 0; x < width; sp += 2, ++vp)
-        {
-          ui32 inf = sp[0];
-          ui32 U_q = sp[1];
-          if (U_q > mmsbp2)
-            return false;
-
-          ui32 v_n;
-          ui32 val = 0;
-          ui32 bit = 0;
-          if (inf & (1 << (4 + bit)))
-          {
-            //get 32 bits of magsgn data
-            ui32 ms_val = frwd_fetch<0xFF>(&magsgn); 
-            ui32 m_n = U_q - ((inf >> (12 + bit)) & 1); // remove e_k
-            frwd_advance(&magsgn, m_n);                 //consume m_n
-
-            val = ms_val << 31;                     // get sign bit
-            v_n = ms_val & ((1 << m_n) - 1);        // keep only m_n bits
-            v_n |= ((inf >> (8 + bit)) & 1) << m_n; // add EMB e_1 as MSB
-            v_n |= 1;                               // add center of bin    
-            //v_n now has 2 * (\mu - 1) + 0.5 with correct sign bit
-            //add 2 to make it 2*\mu+0.5, shift it up to missing MSBs
-            val |= (v_n + 2) << (p - 1);
-          }
-          dp[0] = val;
-
-          v_n = 0;
-          val = 0;
-          bit = 1;
-          if (inf & (1 << (4 + bit)))
-          {
-            //get 32 bits of magsgn data
-            ui32 ms_val = frwd_fetch<0xFF>(&magsgn); 
-            ui32 m_n = U_q - ((inf >> (12 + bit)) & 1); // remove e_k
-            frwd_advance(&magsgn, m_n);                 //consume m_n
-
-            val = ms_val << 31;                      // get sign bit
-            v_n = ms_val & ((1 << m_n) - 1);        // keep only m_n bits
-            v_n |= ((inf >> (8 + bit)) & 1) << m_n; // add EMB e_1 as MSB
-            v_n |= 1;                               // add center of bin    
-            //v_n now has 2 * (\mu - 1) + 0.5 with correct sign bit
-            //add 2 to make it 2*\mu+0.5, shift it up to missing MSBs
-            val |= (v_n + 2) << (p - 1);
-          }
-          dp[stride] = val;
-          vp[0] = prev_v_n | v_n;
-          prev_v_n = 0;
-          ++dp;
-          if (++x >= width)
-          { ++vp; break; }
-
-          val = 0;
-          bit = 2;
-          if (inf & (1 << (4 + bit)))
-          {
-            //get 32 bits of magsgn data
-            ui32 ms_val = frwd_fetch<0xFF>(&magsgn); 
-            ui32 m_n = U_q - ((inf >> (12 + bit)) & 1); // remove e_k
-            frwd_advance(&magsgn, m_n);                 //consume m_n
-
-            val = ms_val << 31;                     // get sign bit
-            v_n = ms_val & ((1 << m_n) - 1);        // keep only m_n bits
-            v_n |= ((inf >> (8 + bit)) & 1) << m_n; // add EMB e_1 as MSB
-            v_n |= 1;                               // add center of bin    
-            //v_n now has 2 * (\mu - 1) + 0.5 with correct sign bit
-            //add 2 to make it 2*\mu+0.5, shift it up to missing MSBs
-            val |= (v_n + 2) << (p - 1);
-          }
-          dp[0] = val;
-
-          v_n = 0;
-          val = 0;
-          bit = 3;
-          if (inf & (1 << (4 + bit)))
-          {
-            //get 32 bits of magsgn data
-            ui32 ms_val = frwd_fetch<0xFF>(&magsgn); 
-            ui32 m_n = U_q - ((inf >> (12 + bit)) & 1); // remove e_k
-            frwd_advance(&magsgn, m_n);                 //consume m_n
-
-            val = ms_val << 31;                     // get sign bit
-            v_n = ms_val & ((1 << m_n) - 1);        // keep only m_n bits
-            v_n |= ((inf >> (8 + bit)) & 1) << m_n; // add EMB e_1 as MSB
-            v_n |= 1;                               // add center of bin    
-            //v_n now has 2 * (\mu - 1) + 0.5 with correct sign bit
-            //add 2 to make it 2*\mu+0.5, shift it up to missing MSBs
-            val |= (v_n + 2) << (p - 1);
-          }
-          dp[stride] = val;
-          prev_v_n = v_n;
-          ++dp;
-          ++x;
-        }
-        vp[0] = prev_v_n;
-
-        for (ui32 y = 2; y < height; y += 2)
-        {
-          ui16 *sp = scratch + (y >> 1) * sstr;
-          ui32 *vp = v_n_scratch;
-          ui32 *dp = decoded_data + y * stride;
-
-          prev_v_n = 0;
-          for (ui32 x = 0; x < width; sp += 2, ++vp)
-          {
-            ui32 inf = sp[0];
-            ui32 u_q = sp[1];
-
-            ui32 gamma = inf & 0xF0; gamma &= gamma - 0x10; //is gamma_q 1?
-            ui32 emax = vp[0] | vp[1];
-            emax = 31 - count_leading_zeros(emax | 2); // emax - 1
-            ui32 kappa = gamma ? emax : 1;
-
-            ui32 U_q = u_q + kappa;
-            if (U_q > mmsbp2)
-              return false;
-
-            ui32 v_n;
-            ui32 val = 0;
-            ui32 bit = 0;
-            if (inf & (1 << (4 + bit)))
-            {
-              //get 32 bits of magsgn data
-              ui32 ms_val = frwd_fetch<0xFF>(&magsgn); 
-              ui32 m_n = U_q - ((inf >> (12 + bit)) & 1); // remove e_k
-              frwd_advance(&magsgn, m_n);                 //consume m_n
-
-              val = ms_val << 31;                     // get sign bit
-              v_n = ms_val & ((1 << m_n) - 1);        // keep only m_n bits
-              v_n |= ((inf >> (8 + bit)) & 1) << m_n; // add EMB e_1 as MSB
-              v_n |= 1;                               // add center of bin    
-              //v_n now has 2 * (\mu - 1) + 0.5 with correct sign bit
-              //add 2 to make it 2*\mu+0.5, shift it up to missing MSBs
-              val |= (v_n + 2) << (p - 1);
-            }
-            dp[0] = val;
-
-            v_n = 0;
-            val = 0;
-            bit = 1;
-            if (inf & (1 << (4 + bit)))
-            {
-              //get 32 bits of magsgn data
-              ui32 ms_val = frwd_fetch<0xFF>(&magsgn); 
-              ui32 m_n = U_q - ((inf >> (12 + bit)) & 1); // remove e_k
-              frwd_advance(&magsgn, m_n);                 //consume m_n
-
-              val = ms_val << 31;                     // get sign bit
-              v_n = ms_val & ((1 << m_n) - 1);        // keep only m_n bits
-              v_n |= ((inf >> (8 + bit)) & 1) << m_n; // add EMB e_1 as MSB
-              v_n |= 1;                               // add center of bin    
-              //v_n now has 2 * (\mu - 1) + 0.5 with correct sign bit
-              //add 2 to make it 2*\mu+0.5, shift it up to missing MSBs
-              val |= (v_n + 2) << (p - 1);
-            }
-            dp[stride] = val;
-            vp[0] = prev_v_n | v_n;
-            prev_v_n = 0;
-            ++dp;
-            if (++x >= width)
-            { ++vp; break; }
-
-            val = 0;
-            bit = 2;
-            if (inf & (1 << (4 + bit)))
-            {
-              //get 32 bits of magsgn data
-              ui32 ms_val = frwd_fetch<0xFF>(&magsgn); 
-              ui32 m_n = U_q - ((inf >> (12 + bit)) & 1); // remove e_k
-              frwd_advance(&magsgn, m_n);                 //consume m_n
-
-              val = ms_val << 31;                     // get sign bit
-              v_n = ms_val & ((1 << m_n) - 1);        // keep only m_n bits
-              v_n |= ((inf >> (8 + bit)) & 1) << m_n; // add EMB e_1 as MSB
-              v_n |= 1;                               // add center of bin    
-              //v_n now has 2 * (\mu - 1) + 0.5 with correct sign bit
-              //add 2 to make it 2*\mu+0.5, shift it up to missing MSBs
-              val |= (v_n + 2) << (p - 1);
-            }
-            dp[0] = val;
-
-            v_n = 0;
-            val = 0;
-            bit = 3;
-            if (inf & (1 << (4 + bit)))
-            {
-              //get 32 bits of magsgn data
-              ui32 ms_val = frwd_fetch<0xFF>(&magsgn); 
-              ui32 m_n = U_q - ((inf >> (12 + bit)) & 1); // remove e_k
-              frwd_advance(&magsgn, m_n);                 //consume m_n
-
-              val = ms_val << 31;                     // get sign bit
-              v_n = ms_val & ((1 << m_n) - 1);        // keep only m_n bits
-              v_n |= ((inf >> (8 + bit)) & 1) << m_n; // add EMB e_1 as MSB
-              v_n |= 1;                               // add center of bin    
-              //v_n now has 2 * (\mu - 1) + 0.5 with correct sign bit
-              //add 2 to make it 2*\mu+0.5, shift it up to missing MSBs
-              val |= (v_n + 2) << (p - 1);
-            }
-            dp[stride] = val;
-            prev_v_n = v_n;
-            ++dp;
-            ++x;
-          }
-          vp[0] = prev_v_n;
-        }
-      }
-
-      if (num_passes > 1)
-      {
-        // We use scratch again, we can divide it into multiple regions
-        // sigma holds all the significant samples, and it cannot
-        // be modified after it is set.  it will be used during the
-        // Magnitude Refinement Pass
-        ui16* const sigma = scratch;
-
-        ui32 mstr = (width + 3u) >> 2;   // divide by 4, since each
-                                         // ui16 contains 4 columns
-        mstr = ((mstr + 2u) + 7u) & ~7u; // multiples of 8
-
-        // We re-arrange quad significance, where each 4 consecutive
-        // bits represent one quad, into column significance, where,
-        // each 4 consequtive bits represent one column of 4 rows
-        {
-          ui32 y;
-          for (y = 0; y < height; y += 4)
-          {
-            ui16* sp = scratch + (y >> 1) * sstr;
-            ui16* dp = sigma + (y >> 2) * mstr;
-            for (ui32 x = 0; x < width; x += 4, sp += 4, ++dp) {
-              ui32 t0 = 0, t1 = 0;
-              t0  = ((sp[0     ] & 0x30u) >> 4)  | ((sp[0     ] & 0xC0u) >> 2);
-              t0 |= ((sp[2     ] & 0x30u) << 4)  | ((sp[2     ] & 0xC0u) << 6);
-              t1  = ((sp[0+sstr] & 0x30u) >> 2)  | ((sp[0+sstr] & 0xC0u)     );
-              t1 |= ((sp[2+sstr] & 0x30u) << 6)  | ((sp[2+sstr] & 0xC0u) << 8);
-              dp[0] = (ui16)(t0 | t1);
-            }
-            dp[0] = 0; // set an extra entry on the right with 0
-          }
-          {
-            // reset one row after the codeblock
-            ui16* dp = sigma + (y >> 2) * mstr;
-            for (ui32 x = 0; x < width; x += 4, ++dp)
-              dp[0] = 0;
-            dp[0] = 0; // set an extra entry on the right with 0
-          }
-        }
-
-        // We perform Significance Propagation Pass here
-        {
-          // This stores significance information of the previous
-          // 4 rows.  Significance information in this array includes
-          // all signicant samples in bitplane p - 1; that is,
-          // significant samples for bitplane p (discovered during the
-          // cleanup pass and stored in sigma) and samples that have recently
-          // became significant (during the SPP) in bitplane p-1.
-          // We store enough for the widest row, containing 1024 columns,
-          // which is equivalent to 256 of ui16, since each stores 4 columns.
-          // We add an extra 8 entries, just in case we need more
-          ui16 prev_row_sig[256 + 8] = {0}; // 528 Bytes
-
-          frwd_struct sigprop;
-          frwd_init<0>(&sigprop, coded_data + lengths1, (int)lengths2);
-
-          for (ui32 y = 0; y < height; y += 4)
-          {
-            ui32 pattern = 0xFFFFu; // a pattern needed samples
-            if (height - y < 4) {
-              pattern = 0x7777u;
-              if (height - y < 3) {
-                pattern = 0x3333u;
-                if (height - y < 2)
-                  pattern = 0x1111u;
-              }
-            }
-
-            // prev holds sign. info. for the previous quad, together
-            // with the rows on top of it and below it.
-            ui32 prev = 0;
-            ui16 *prev_sig = prev_row_sig;
-            ui16 *cur_sig = sigma + (y >> 2) * mstr;
-            ui32 *dpp = decoded_data + y * stride;
-            for (ui32 x = 0; x < width; x += 4, ++cur_sig, ++prev_sig)
-            {
-              // only rows and columns inside the stripe are included
-              si32 s = (si32)x + 4 - (si32)width;
-              s = ojph_max(s, 0);
-              pattern = pattern >> (s * 4);
-
-              // We first find locations that need to be tested (potential
-              // SPP members); these location will end up in mbr
-              // In each iteration, we produce 16 bits because cwd can have
-              // up to 16 bits of significance information, followed by the
-              // corresponding 16 bits of sign information; therefore, it is
-              // sufficient to fetch 32 bit data per loop.
-
-              // Althougth we are interested in 16 bits only, we load 32 bits.
-              // For the 16 bits we are producing, we need the next 4 bits --
-              // We need data for at least 5 columns out of 8.
-              // Therefore loading 32 bits is easier than loading 16 bits
-              // twice.
-              ui32 ps = *(ui32*)prev_sig;
-              ui32 ns = *(ui32*)(cur_sig + mstr);
-              ui32 u = (ps & 0x88888888) >> 3; // the row on top
-              if (!stripe_causal)
-                u |= (ns & 0x11111111) << 3;   // the row below
-
-              ui32 cs = *(ui32*)cur_sig;
-              // vertical integration
-              ui32 mbr =  cs;                // this sig. info.
-              mbr |= (cs & 0x77777777) << 1; //above neighbors
-              mbr |= (cs & 0xEEEEEEEE) >> 1; //below neighbors
-              mbr |= u;
-              // horizontal integration
-              ui32 t = mbr;
-              mbr |= t << 4;      // neighbors on the left
-              mbr |= t >> 4;      // neighbors on the right
-              mbr |= prev >> 12;  // significance of previous group
-
-              // remove outside samples, and already significant samples
-              mbr &= pattern;
-              mbr &= ~cs;
-
-              // find samples that become significant during the SPP
-              ui32 new_sig = mbr;
-              if (new_sig)
-              {
-                ui32 cwd = frwd_fetch<0>(&sigprop);
-
-                ui32 cnt = 0;
-                ui32 col_mask = 0xFu;
-                ui32 inv_sig = ~cs & pattern;
-                for (int i = 0; i < 16; i += 4, col_mask <<= 4)
-                {
-                  if ((col_mask & new_sig) == 0)
-                    continue;
-
-                  //scan one column
-                  ui32 sample_mask = 0x1111u & col_mask;
-                  if (new_sig & sample_mask)
-                  {
-                    new_sig &= ~sample_mask;
-                    if (cwd & 1)
-                    {
-                      ui32 t = 0x33u << i;
-                      new_sig |= t & inv_sig;
-                    }
-                    cwd >>= 1; ++cnt;
-                  }
-
-                  sample_mask <<= 1;
-                  if (new_sig & sample_mask)
-                  {
-                    new_sig &= ~sample_mask;
-                    if (cwd & 1)
-                    {
-                      ui32 t = 0x76u << i;
-                      new_sig |= t & inv_sig;
-                    }
-                    cwd >>= 1; ++cnt;
-                  }
-
-                  sample_mask <<= 1;
-                  if (new_sig & sample_mask)
-                  {
-                    new_sig &= ~sample_mask;
-                    if (cwd & 1)
-                    {
-                      ui32 t = 0xECu << i;
-                      new_sig |= t & inv_sig;
-                    }
-                    cwd >>= 1; ++cnt;
-                  }
-
-                  sample_mask <<= 1;
-                  if (new_sig & sample_mask)
-                  {
-                    new_sig &= ~sample_mask;
-                    if (cwd & 1)
-                    {
-                      ui32 t = 0xC8u << i;
-                      new_sig |= t & inv_sig;
-                    }
-                    cwd >>= 1; ++cnt;
-                  }
-                }
-
-                if (new_sig)
-                {
-                  // new_sig has newly-discovered sig. samples during SPP
-                  // find the signs and update decoded_data
-                  ui32 *dp = dpp + x;
-                  ui32 val = 3u << (p - 2);
-                  col_mask = 0xFu;
-                  for (int i = 0; i < 4; ++i, ++dp, col_mask <<= 4)
-                  {
-                    if ((col_mask & new_sig) == 0)
-                      continue;
-
-                    //scan 4 signs
-                    ui32 sample_mask = 0x1111u & col_mask;
-                    if (new_sig & sample_mask)
-                    {
-                      assert(dp[0] == 0);
-                      dp[0] = (cwd << 31) | val;
-                      cwd >>= 1; ++cnt;
-                    }
-
-                    sample_mask += sample_mask;
-                    if (new_sig & sample_mask)
-                    {
-                      assert(dp[stride] == 0);
-                      dp[stride] = (cwd << 31) | val;
-                      cwd >>= 1; ++cnt;
-                    }
-
-                    sample_mask += sample_mask;
-                    if (new_sig & sample_mask)
-                    {
-                      assert(dp[2 * stride] == 0);
-                      dp[2 * stride] = (cwd << 31) | val;
-                      cwd >>= 1; ++cnt;
-                    }
-
-                    sample_mask += sample_mask;
-                    if (new_sig & sample_mask)
-                    {
-                      assert(dp[3 * stride] == 0);
-                      dp[3 * stride] = (cwd << 31) | val;
-                      cwd >>= 1; ++cnt;
-                    }
-                  }
-                }
-                frwd_advance(&sigprop, cnt);
-              }
-
-              new_sig |= cs;
-              *prev_sig = (ui16)(new_sig);
-
-              // vertical integration for the new sig. info.
-              t = new_sig;
-              new_sig |= (t & 0x7777) << 1; //above neighbors
-              new_sig |= (t & 0xEEEE) >> 1; //below neighbors
-              // add sig. info. from the row on top and below
-              prev = new_sig | u;
-              // we need only the bits in 0xF000
-              prev &= 0xF000;
-            }
-          }
-        }
-
-        // We perform Magnitude Refinement Pass here
-        if (num_passes > 2)
-        {
-          rev_struct magref;
-          rev_init_mrp(&magref, coded_data, (int)lengths1, (int)lengths2);
-
-          for (ui32 y = 0; y < height; y += 4)
-          {
-            ui32 *cur_sig = (ui32*)(sigma + (y >> 2) * mstr);
-            ui32 *dpp = decoded_data + y * stride;
-            ui32 half = 1 << (p - 2);
-            for (ui32 i = 0; i < width; i += 8)
-            {
-              //Process one entry from sigma array at a time
-              // Each nibble (4 bits) in the sigma array represents 4 rows,
-              // and the 32 bits contain 8 columns
-              ui32 cwd = rev_fetch_mrp(&magref); // get 32 bit data
-              ui32 sig = *cur_sig++; // 32 bit that will be processed now
-              ui32 col_mask = 0xFu;  // a mask for a column in sig
-              if (sig) // if any of the 32 bits are set
-              {
-                for (int j = 0; j < 8; ++j) //one column at a time
-                {
-                  if (sig & col_mask) // lowest nibble
-                  {
-                    ui32 *dp = dpp + i + j; // next column in decoded samples
-                    ui32 sample_mask = 0x11111111u & col_mask; //LSB
-
-                    for (int k = 0; k < 4; ++k) {
-                      if (sig & sample_mask) //if LSB is set
-                      {
-                        assert(dp[0] != 0); // decoded value cannot be zero
-                        assert((dp[0] & half) == 0); // no half
-                        ui32 sym = cwd & 1;          // get it value
-                        sym = (1 - sym) << (p - 1); // previous center of bin
-                        sym |= half;            // put half the center of bin
-                        dp[0] ^= sym;    // remove old bin center and put new
-                        cwd >>= 1;       // consume word
-                      }
-                      sample_mask += sample_mask; //next row
-                      dp += stride; // next samples row
-                    }
-                  }
-                  col_mask <<= 4; //next column
-                }
-              }
-              // consume data according to the number of bits set
-              rev_advance_mrp(&magref, population_count(sig));
-            }
-          }
-        }
-      }
-      return true;
-    }
 
     //************************************************************************/
     /** @brief Decodes one codeblock, processing the cleanup, siginificance
